requirements.txt
PyQt6>=6.10.0
PyQt6-WebEngine>=6.10.0
python-barcode==0.15.1
Pillow==10.1.0
pyzbar==0.1.9
numpy>=1.21.0
python-dotenv>=1.0.0

